requirements.txt
streamlit==1.31.0
pandas==2.2.0
plotly==5.18.0
wordcloud==1.9.3
matplotlib==3.8.2
seaborn==0.13.1
nltk==3.8.1
textblob==0.18.0
pillow==10.2.0
openpyxl==3.1.2
pyinstaller==6.3.0
boto3>=1.34.0
python-dotenv>=1.0.0
python-pptx>=0.6.21
fpdf2>=2.7.0
